Implements suport for server-side loading of source files.
ASPX files and other formats will be interpreted by server and actual
source code of the files are not possible to be reached from JavaScript
(clientside) only. The approach created implements a direct method to
allow collecting file contents directly from server with the following
limitations:
- files must be under AppDomain directory tree
- files must be of a given set of acceptable extensions

NOTICE: Do not add config to the list of allowed extensions or a
malicious user would be able to access Web.config.